feat: allow to define common labels in k8s-image-swapper chart
This commit add a new parameter `commonLabels` in the `values.yaml`
that allows to define a set of labels that will be set on all the
chart resources. They will not be added to the selectors.

This is particularly useful when trying to "tag" the resources and
indicate their associated team etc